pyproject.toml
[tool.black]
line-length = 88
skip-string-normalization = true
target-version = ["py36", "py37", "py38"]

[tool.poetry]
name = "django-scoped-rbac"
version = "0.7.0"
description = "A rich and flexible Django application for role-based access control within distinct access control scopes supporting Django Rest Framework."
authors = ["David Charboneau <david@adadabase.com>"]
license = "MIT"
repository = "https://gitlab.com/openteams/django-scoped-rbac"
packages = [
  { include = "scoped_rbac", from = "src"},
  { include = "celpy", from = "cel-python/src"},
  { include = "xlate", from = "cel-python/src"},
]

[tool.poetry.dependencies]
python = "^3.7.6"
django = "^2.2.8"
djangorestframework = "^3.11.0"
django-rest-framework-condition = "^0.1.0"
scrudful = "^0.1.0"
drf-extensions = "^0.6.0"
django-jsonfield-backport = "^1.0.3"
jmespath = "^0.10.0"
lark-parser = "^0.8.5"
python-dateutil = "^2.8.1"
pyyaml = "^5.3.1"
requests = "^2.25.0"
urllib3 = "^1.26.1"
Babel = "^2.9.0"

[tool.poetry.dev-dependencies]
astor = { git="https://github.com/Quansight/astor.git", branch="issue-174" }
black = "19.3b0"
coverage = "^5.3.1"
factory-boy = "^3.2.0"
flake8 = "^3.8.4"
pytest = "^6.2.1"
pytest-cov = "^2.10.1"
pytest-django = "^4.1.0"
# safetydance = { git="https://gitlab.com/openteams/safetydance.git", branch="python38" }
safetydance-django = { git = "https://gitlab.com/openteams/safetydance_django.git" }
Sphinx = "^2.3.1"
sphinxcontrib-django = "^0.5.1"
sphinx-autobuild = "^2020.9.1"
sphinx-rtd-theme = "^0.5.0"
# safetydance-test = "^0.1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry2conda]
name = "django-scoped-rbac"
